+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Switch from `unzipper` to `yauzl` internally
+
 ## [0.1.0] - 2025-03-07
 
 ### Added